Convert utils filte to TS, lint fixes, read config dir refactor
