+++ packages/luis/package.json
@@ -8,6 +8,7 @@
     "@oclif/config": "^1.13.3",
     "@types/sinon": "^7.5.0",
     "azure-cognitiveservices-luis-authoring": "2.1.1",
+    "fs-extra": "^8.1.0",
     "ms-rest": "2.5.3",
     "tslib": "^1.10.0"
   },
@@ -46,7 +47,7 @@
   "license": "MIT",
   "oclif": {
     "commands": "./lib/commands",
-    "bin": "oclif-example",
+    "bin": "bf",
     "devPlugins": [
       "@oclif/plugin-help"
     ]
